requirements.txt
nextcord
python-dotenv
dnspython
nextcord-ext-activities
jishaku
